requirements.txt
requests
python-dotenv
wakeonlan